# packages/sdk/openapi/v1.yaml
openapi: 3.0.3
info:
  title: Synaro Public API
  version: "1.0.0"
  description: |
    Bearer API-key surface at `/api/v1`. Mint keys in the Synaro dashboard (Settings → API keys).
    TypeScript clients should prefer `@synaro/sdk`. Agent **write** bodies use camelCase
    (`systemPrompt`, `toolMode`, …); the server also accepts snake_case aliases.
servers:
  - url: https://synaro.tech
  - url: http://localhost:3000
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: sk_live
...
paths:
...
  /api/v1/agents/{agentId}/memory:
    get:
      summary: List agent memory
      parameters:
        - name: agentId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
    delete:
      summary: Clear agent memory
      parameters:
        - name: agentId
          in: path
          required: true
          schema: { type: string }
      responses:
        "204": { description: Cleared }
  /api/v1/agents/{agentId}/memory/{key}:
    put:
      summary: Upsert memory entry
      parameters:
        - name: agentId
          in: path
          required: true
          schema: { type: string }
        - name: key
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
    delete:
      summary: Delete memory entry
      parameters:
        - name: agentId
          in: path
          required: true
          schema: { type: string }
        - name: key
          in: path
          required: true
          schema: { type: string }
      responses:
        "204": { description: Deleted }
...